V 3.0.1
Version 3.0.1
05/09/20:
Moved “helps privacy causes” to the Additionally Address section as this is both A) Relative and B) Doesn’t directly affect a service in any meaningful way that should affect a systematic scoring method. Moving it to additionally address still allows discussion on the matter without unfairly docking points for forced cooperation.