Elaborating on our affiliate minimum requirement standard
Some people have asked why we have our standards set to 4.20/5. This is simply a standard we set for ourselves to be able to offer the users a choice of an affiliate plan to support us. We don’t want to open affiliate plans and be associated with any services we don’t personally WANT to associate with. Our previous limit was 4.0/5 but there were several services we didn’t feel comfortable using affiliate links with. Raising this personal requirement has 0 impact on the VPNs being tested, their scores are still the same, the only people this effects is ourselves, as we are now unable to make money with services under that requirement.